+++ packages/terra-core-docs/src/terra-dev-site/test/card/CardVisuallyHiddenText.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Card from 'terra-card';
+
+const CardVisuallyHiddenText = () => (
+  <div>
+    <Card visuallyHiddenText="This is a Hello World Card Introduction">Hello World!!</Card>
+  </div>
+);
+
+export default CardVisuallyHiddenText;