+++ packages/terra-core-docs/src/terra-dev-site/test/demographics-banner/GestationalAgeDemographicsBannerNoLabel.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import DemographicsBanner from 'terra-demographics-banner';
+
+export default () => <DemographicsBanner id="gestational-age-no-label" gestationalAge="April 5, 2016" />;